maul

+++ routes/web.php
@@ -31,57 +31,59 @@ use App\Http\Controllers\CMS\PortfolioCategoryController;
 Route::get('/', [WelcomeController::class, 'index']);
 
 //blog
-Route::get('/blog/{slug}', [BlogPostController::class, 'findBlogPost']);
-Route::get('/blogs', [BlogPostController::class, 'index']);
+Route::get('/blog/{slug}', [BlogPostController::class, 'singleBlog'])->name('blog');
+Route::get('/blogs', [BlogPostController::class, 'allBlogs']);
 
 //comments
-Route::post('/blog/{slug}/comment', [CommentController::class, 'store']);
+Route::post('/blog/{slug}/comment', [CommentController::class, 'store'])->name('blog.comment');
 
 //contact us
-Route::post('/contact',[ContactController::class ,'store']);
+Route::post('/contact',[ContactController::class ,'store'])->name('contact');
 
 //Route group for authenticated users only
 Route::group(['middleware' => ['auth']], function()
 {
-    Route::get('/dashboard', [DashboardController::class, 'index']);
-    Route::get('/profile', [SettingsController::class, 'index']);
-    Route::get('/cms', [DashboardController::class, 'cmsView']);
+    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+    Route::get('/profile', [SettingsController::class, 'index'])->name('profile');
+    Route::get('/cms', [DashboardController::class, 'cmsView'])->name('cms');
 
-    Route::get('/logout',[LoginController::class ,'logout']);
-    Route::put('/settings/profile',[SettingsController::class ,'updateProfile']);
-    Route::put('/settings/password',[SettingsController::class ,'updatePassword']);
+    Route::get('/logout',[LoginController::class ,'logout'])->name('logout');
+    Route::put('/settings/profile',[SettingsController::class ,'updateProfile'])->name('settings.profile');
+    Route::put('/settings/password',[SettingsController::class ,'updatePassword'])->name('settings.password');
 
     //blog post
-    Route::get('/admin/blog/{slug}', [BlogPostController::class, 'adminSingleBlogPost']);
-    Route::get('/admin/blogs', [BlogPostController::class, 'adminAllBlogPosts']);
-    Route::get('/admin/blog',[BlogPostController::class ,'createView']);
-    Route::post('/admin/blog',[BlogPostController::class ,'store']);
-    Route::put('/blog/{slug}', [BlogPostController::class, 'update']);
-    Route::delete('/blog/{slug}', [BlogPostController::class, 'destroy']);
+    Route::get('/admin/blog',[BlogPostController::class ,'createView'])->name('admin.blog'); 
+    Route::get('/admin/blog/{slug}', [BlogPostController::class, 'adminSingleBlogPost'])->name('admin.single_blog');
+    Route::get('/admin/blogs', [BlogPostController::class, 'adminAllBlogPosts'])->name('admin.blogs');
+    Route::post('/admin/blog/create',[BlogPostController::class ,'store'])->name('admin.blog.create');
+    Route::put('/admin/blog/update/{slug}',[BlogPostController::class ,'update'])->name('admin.blog.update');
+    Route::get('/admin/blog/delete/{slug}', [BlogPostController::class, 'destroy'])->name('admin.blog.delete');
 
     //comments
-    Route::delete('/comment/{id}', [CommentController::class, 'destroy']);
+    Route::get('/admin/comment/delete/{id}', [CommentController::class, 'destroy'])->name('admin.comment.delete');
 
     //CMS
     //banner
-    Route::put('/banner/{id}', [BannerController::class, 'update']);
+    Route::put('/admin/banner/update/{id}', [BannerController::class, 'update'])->name('admin.banner.update');
+    Route::put('/admin/banner/update_image/{id}', [BannerController::class, 'updateImage'])->name('admin.banner.update_image');
     //about
-    Route::put('/banner/{id}', [AboutController::class, 'update']);
+    Route::put('/admin/about/about/{id}', [AboutController::class, 'update'])->name('admin.about.update');
+    Route::put('/admin/about/about_image/{id}', [AboutController::class, 'updateImage'])->name('admin.about.update_image');
+    
 
     //portfolio category
-    Route::post('/portfolio_category',[PortfolioCategoryController::class ,'store']);
-    Route::put('/portfolio_category/{id}', [PortfolioCategoryController::class, 'update']);
-    Route::delete('/portfolio_category/{id}', [PortfolioCategoryController::class, 'destroy']);
+    Route::post('/admin/portfolio_category/create',[PortfolioCategoryController::class ,'store'])->name('admin.portfolio_category.create');
+    Route::put('/admin/portfolio_category/update/{id}', [PortfolioCategoryController::class, 'update'])->name('admin.portfolio_category.update');
+    Route::delete('/admin/portfolio_category/delete/{id}', [PortfolioCategoryController::class, 'destroy'])->name('admin.portfolio_category.delete');
 
     //portfolio 
-    Route::post('/portfolio',[PortfolioController::class ,'store']);
-    Route::put('/portfolio/{id}', [PortfolioController::class, 'update']);
-    Route::delete('/portfolio/{id}', [PortfolioController::class, 'destroy']);
+    Route::post('/admin/portfolio/create',[PortfolioController::class ,'store'])->name('admin.portfolio.create');
+    Route::put('/admin/portfolio/update/{id}', [PortfolioController::class, 'update'])->name('admin.portfolio.update');
+    Route::put('/admin/portfolio/update_image/{id}', [PortfolioController::class, 'updateImage'])->name('admin.portfolio.update_image');
+    Route::delete('/admin/portfolio/delete/{id}', [PortfolioController::class, 'destroy'])->name('admin.portfolio.delete');
 
     //Social media 
-    Route::post('/social_media',[SocialMediaController::class ,'store']);
-    Route::put('/social_media/{id}', [SocialMediaController::class, 'update']);
-    Route::delete('/social_media/{id}', [SocialMediaController::class, 'destroy']);
+    Route::put('/admin/social_media/update/{id}', [SocialMediaController::class, 'update'])->name('admin.social_media.update');
 });
 
 //Route group for authenticated GUEST only
